Added new file, moved algorithms there
