+++ math/src/fft/mod.rs
@@ -1,4 +1,5 @@
 pub mod errors;
 pub mod fft_cooley_tukey;
+pub mod fft_iterative;
 mod helpers;
 pub mod operations;
